+++ tutorials/cpp/fixeddecimal.cpp
@@ -35,7 +35,7 @@ int main() {
         return 1;
     }
 
-    decimal = FixedDecimal::from_double_with_floating_precision(100.01).ok().value();
+    decimal = FixedDecimal::from_double_with_round_trip_precision(100.01).ok().value();
     out = fdf->format(*decimal.get());
     std::cout << "Formatted float value is " << out << std::endl;
     if (out != "১০০.০১") {
@@ -87,14 +87,14 @@ int main() {
 
     fdf = FixedDecimalFormatter::create_with_manual_data("+", "", "-", "", "/", "_", 4, 2, 4, digits, FixedDecimalGroupingStrategy::Auto).ok().value();
 
-    decimal = FixedDecimal::from_double_with_floating_precision(123456.8901).ok().value();
+    decimal = FixedDecimal::from_double_with_round_trip_precision(123456.8901).ok().value();
     out = fdf->format(*decimal.get());
     std::cout << "Formatted float value for custom numeric system is " << out << std::endl;
     if (out != "bcdefg/ijab") {
         std::cout << "Output does not match expected output" << std::endl;
         return 1;
     }
-    decimal = FixedDecimal::from_double_with_floating_precision(123451234567.8901).ok().value();
+    decimal = FixedDecimal::from_double_with_round_trip_precision(123451234567.8901).ok().value();
     out = fdf->format(*decimal.get());
     std::cout << "Formatted float value for custom numeric system is " << out << std::endl;
     if (out != "bc_de_fb_cd_efgh/ijab") {
@@ -107,7 +107,7 @@ int main() {
     fdf = FixedDecimalFormatter::create_with_grouping_strategy(
         *dp.get(), *locale.get(), FixedDecimalGroupingStrategy::Auto).ok().value();
 
-    decimal = FixedDecimal::from_double_with_floating_precision(123456.8901).ok().value();
+    decimal = FixedDecimal::from_double_with_round_trip_precision(123456.8901).ok().value();
     out = fdf->format(*decimal.get());
     std::cout << "Formatted value is " << out << std::endl;
     if (out != "๑๒๓,๔๕๖.๘๙๐๑") {